\name{TestWrapper}
\alias{TestWrapper}
\title{ Test Wrapper Function for ExaGeoStatR interface to the CalcMatAvgR() function  }
\description{
  \code{TestWrapper} demonstrates the R / C language interface
}

\usage{
TestWrapper()
}

\arguments{ None. }

\value{ Returns nothing. }


\author{ Sameh Abdulah, Postdoctoral Fellow, Extreme Computing and Research Center, KAUST \email{sameh.abdulah@kaust.edu.sa}}

\examples{
#
# inline call at R command line or in another R function:
# Provides function arguments for and calls the R CalcMatAvgR() function.
#
#   TestWrapper()
}
\keyword{programming}

